Rb 37 start game with gamedefinition name

See merge request robolucha/robolucha-runner!13

+++ src/main/java/com/robolucha/game/action/AddOnListenEventAction.java
@@ -24,7 +24,7 @@ public class AddOnListenEventAction implements GameAction {
 		this.event = new OnListenEvent(source, message);
 	}
 
-	public void run(LinkedHashMap<Long, LuchadorRunner> runners,
+	public void run(LinkedHashMap<Integer, LuchadorRunner> runners,
 			LuchadorRunner runner) throws Exception {
 
 		if (logger.isDebugEnabled()) {
